Handle array and function indexing cleanly
If it's an array indexing, it's always correct, e.g. 'u' in the following is
always an array:

u(1, :)

If it's a function call, it can also be an array indexing, based on the
function/array symbol type, e.g. 'u' in the following can be either a function
or an array:

u(1, 2)
u(1, i)

That has to be decided later after parsing is done.

Tests slowed down by this change from 0.77s to 1.12s.