Allocate strings correctly
One should not use C casts like that. In this case the const pointer was cast
away and then when the visitor goes out of scope, the `std::string` and thus
the pointer gets deallocated and we have a dangling pointer in the ASR. Strings
must be allocated using our Allocator, for example via the Str convenience
class.