Alternative implementation of statement functions
We have made the parse more general and allow to mix declarations and
statements. The statement function then gets parsed as a regular assignment.
The ast2asr pass would then figure out whether it is an assignment or a
statement function and create the ASR accordingly.

This approach actually seems to reduce the number of grammar conflicts. But it
requires to modify the ast2asr more extensively, which was not done in this MR.

Now this:

    program Axxxxxxxxxx
    CHARACTER(len=4) :: T
    !
    ROTX(X,Y)=CO*X-SI*Y+X01
    ROTY(X,Y)=SI*X+CO*Y+Y01
    DATA SPACE/0.08/
    integer :: c2
    a(x) = 3
    integer :: c
    !
    TL=0.1            ! TICK LENGTH
    end

Parses to:

(TranslationUnit [(Program Axxxxxxxxxx () [] [] [] [] [])])

Because we had to not include declarations and body, so that the rest of the
code compiles.