C-preprocessor: Make defined(Y) && (3 <= Y) work
Previously if Y was not defined, it would still try to evaluate it in the 3 <=
Y part and give an error when Y was not defined. It seems however that it is
not an error to try to evaluate a macro that does not exist, so we have made
that change. Now the above works.

Tests added.