Return 0 from a main() function
This is only temporary, eventually we should make it mandatory for the user to
do that explicitly (with proper declarations and such), but for now this should
work.

As a result of this, the LLVM now declares the main() function to return i64,
and we always return 0. That should return 0 exit code on all platforms.