Rename the @jit decorator to @lpython
Added inital docstring explaining the motivation. The term `jit` is not
the right term for this decorator because we are not using JIT (=Just In
Time) compilation, but rather AOT (=Ahead of Time) compilation, at
runtime from CPython. The term `lpython` is better, as it conveys that
we are using LPython to compile the function (ahead of time). It is also
consistent with the @cpython decorator.