RTL: fix an issue with abs()
Previously the abs() was somehow returning a negative number, because we
called it with an int8_t argument, but it was expecting "int" argument.

The other bug was that "num" in the for loop stayed at -1 forever. We
fixed that by using the condition "num > 0". We also switched to a while
loop which make it more readable.