Consolidate cpython_py and cpython
The two tests seem to be the same. The `c_py` requires Python to be
installed which otherwise is not required. But the `cpython_py` just
requires Python to be installed, which it is anyway for the `cpython`
backend, so the two labels are effectively the same. So we make them the
same.